+++ src/Commands/ImportAccountData.php
@@ -4,9 +4,9 @@ namespace Inovector\Mixpost\Commands;
 
 use Illuminate\Console\Command;
 use Inovector\Mixpost\Concerns\AccountsOption;
-use Inovector\Mixpost\Jobs\ImportFacebookInsightsJob;
-use Inovector\Mixpost\Jobs\ImportMastodonPostsJob;
-use Inovector\Mixpost\Jobs\ImportTwitterPostsJob;
+use Inovector\Mixpost\SocialProviders\Mastodon\Jobs\ImportMastodonPostsJob;
+use Inovector\Mixpost\SocialProviders\Meta\Jobs\ImportFacebookInsightsJob;
+use Inovector\Mixpost\SocialProviders\Twitter\Jobs\ImportTwitterPostsJob;
 
 class ImportAccountData extends Command
 {
