+++ src/SocialProviders/Mastodon/Jobs/ImportMastodonFollowersJob.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inovector\Mixpost\Jobs;
+namespace Inovector\Mixpost\SocialProviders\Mastodon\Jobs;
 
 use Carbon\Carbon;
 use Illuminate\Bus\Batchable;
@@ -14,10 +14,10 @@ use Inovector\Mixpost\Concerns\Job\SocialProviderJobFail;
 use Inovector\Mixpost\Concerns\UsesSocialProviderManager;
 use Inovector\Mixpost\Models\Account;
 use Inovector\Mixpost\Models\Audience;
-use Inovector\Mixpost\SocialProviders\Meta\FacebookGroupProvider;
+use Inovector\Mixpost\SocialProviders\Mastodon\MastodonProvider;
 use Inovector\Mixpost\Support\SocialProviderResponse;
 
-class ImportFacebookGroupMembersJob implements ShouldQueue
+class ImportMastodonFollowersJob implements ShouldQueue
 {
     use Batchable, Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
@@ -43,10 +43,10 @@ class ImportFacebookGroupMembersJob implements ShouldQueue
         }
 
         /**
-         * @see FacebookGroupProvider
+         * @see MastodonProvider
          * @var SocialProviderResponse $response
          */
-        $response = $this->connectProvider($this->account)->getGroupMetrics();
+        $response = $this->connectProvider($this->account)->getAccountMetrics();
 
         if ($response->hasExceededRateLimit()) {
             $this->storeRateLimitExceeded($response->retryAfter(), $response->isAppLevel());
@@ -67,9 +67,9 @@ class ImportFacebookGroupMembersJob implements ShouldQueue
 
         Audience::updateOrCreate([
             'account_id' => $this->account->id,
-            'date' => Carbon::today('UTC')->toDateString()
+            'date' => Carbon::today('UTC')
         ], [
-            'total' => $response->member_count ?? 0
+            'total' => $response->followers_count ?? 0
         ]);
     }
 }
