+++ src/SocialProviders/Mastodon/Jobs/ImportMastodonPostsJob.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inovector\Mixpost\Jobs;
+namespace Inovector\Mixpost\SocialProviders\Mastodon\Jobs;
 
 use Carbon\Carbon;
 use Illuminate\Bus\Batchable;
@@ -10,17 +10,15 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Str;
 use Inovector\Mixpost\Concerns\Job\HasSocialProviderJobRateLimit;
 use Inovector\Mixpost\Concerns\Job\SocialProviderJobFail;
 use Inovector\Mixpost\Concerns\UsesSocialProviderManager;
-use Inovector\Mixpost\Enums\FacebookInsightType;
 use Inovector\Mixpost\Models\Account;
-use Inovector\Mixpost\Models\FacebookInsight;
-use Inovector\Mixpost\SocialProviders\Meta\FacebookPageProvider;
+use Inovector\Mixpost\Models\ImportedPost;
+use Inovector\Mixpost\SocialProviders\Mastodon\MastodonProvider;
 use Inovector\Mixpost\Support\SocialProviderResponse;
 
-class ImportFacebookInsightsJob implements ShouldQueue
+class ImportMastodonPostsJob implements ShouldQueue
 {
     use Batchable, Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
@@ -31,10 +29,12 @@ class ImportFacebookInsightsJob implements ShouldQueue
     public $deleteWhenMissingModels = true;
 
     public Account $account;
+    public array $params;
 
-    public function __construct(Account $account)
+    public function __construct(Account $account, array $params = [])
     {
         $this->account = $account;
+        $this->params = $params;
     }
 
     public function handle(): void
@@ -46,10 +46,10 @@ class ImportFacebookInsightsJob implements ShouldQueue
         }
 
         /**
-         * @see FacebookPageProvider
+         * @see MastodonProvider
          * @var SocialProviderResponse $response
          */
-        $response = $this->connectProvider($this->account)->getPageInsights();
+        $response = $this->connectProvider($this->account)->getUserStatuses($this->account->provider_id, $this->params['max_id'] ?? '');
 
         if ($response->hasExceededRateLimit()) {
             $this->storeRateLimitExceeded($response->retryAfter(), $response->isAppLevel());
@@ -68,24 +68,39 @@ class ImportFacebookInsightsJob implements ShouldQueue
             return;
         }
 
-        $insights = $response->context()['data'];
+        $posts = $this->filterPosts($response->data);
 
-        foreach ($insights as $insight) {
-            $this->importInsights(FacebookInsightType::fromName(Str::upper($insight['name'])), $insight['values']);
+        $this->import($posts);
+
+        // If more than 40(limit of page items), there is a probability that there are others.
+        if (count($posts) >= 40) {
+            ImportMastodonPostsJob::dispatch($this->account, ['max_id' => Arr::last($posts)['id']])->delay(60 * 15);
         }
     }
 
-    protected function importInsights(FacebookInsightType $type, array $items): void
+    protected function filterPosts(array $data): array
     {
-        $data = Arr::map($items, function ($item) use ($type) {
+        return Arr::where($data, function ($item) {
+            return Carbon::parse($item['created_at'], 'UTC')->greaterThan(Carbon::now('UTC')->subDays(90));
+        });
+    }
+
+    protected function import(array $items): void
+    {
+        $data = Arr::map($items, function ($item) {
             return [
                 'account_id' => $this->account->id,
-                'type' => $type,
-                'date' => Carbon::parse($item['end_time'], 'UTC')->toDateString(),
-                'value' => $item['value'],
+                'provider_post_id' => $item['id'],
+                'content' => json_encode(['text' => $item['content']]),
+                'metrics' => json_encode([
+                    'replies' => $item['replies_count'],
+                    'reblogs' => $item['reblogs_count'],
+                    'favourites' => $item['favourites_count'],
+                ]),
+                'created_at' => Carbon::parse($item['created_at'], 'UTC')->toDateString()
             ];
         });
 
-        FacebookInsight::upsert($data, ['account_id', 'type', 'date'], ['value']);
+        ImportedPost::upsert($data, ['account_id', 'provider_post_id'], ['content', 'metrics']);
     }
 }
