+++ src/SocialProviders/Twitter/Jobs/ProcessTwitterMetricsJob.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inovector\Mixpost\Jobs;
+namespace Inovector\Mixpost\SocialProviders\Twitter\Jobs;
 
 use Illuminate\Bus\Batchable;
 use Illuminate\Bus\Queueable;
@@ -13,7 +13,7 @@ use Inovector\Mixpost\Models\Account;
 use Inovector\Mixpost\Models\ImportedPost;
 use Inovector\Mixpost\Models\Metric;
 
-class ProcessMastodonMetricsJob implements ShouldQueue
+class ProcessTwitterMetricsJob implements ShouldQueue
 {
     use Batchable, Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
@@ -29,9 +29,10 @@ class ProcessMastodonMetricsJob implements ShouldQueue
     public function handle()
     {
         $items = ImportedPost::select('created_at',
+            DB::raw('SUM(JSON_EXTRACT(metrics, "$.likes")) as likes'),
             DB::raw('SUM(JSON_EXTRACT(metrics, "$.replies")) as replies'),
-            DB::raw('SUM(JSON_EXTRACT(metrics, "$.reblogs")) as reblogs'),
-            DB::raw('SUM(JSON_EXTRACT(metrics, "$.favourites")) as favourites'))
+            DB::raw('SUM(JSON_EXTRACT(metrics, "$.retweets")) as retweets'),
+            DB::raw('SUM(JSON_EXTRACT(metrics, "$.impressions")) as impressions'))
             ->where('account_id', $this->account->id)
             ->groupBy('created_at')
             ->cursor();
@@ -41,9 +42,10 @@ class ProcessMastodonMetricsJob implements ShouldQueue
                 'account_id' => $this->account->id,
                 'date' => $item->created_at,
                 'data' => json_encode([
+                    'likes' => $item->likes,
                     'replies' => $item->replies,
-                    'reblogs' => $item->reblogs,
-                    'favourites' => $item->favourites,
+                    'retweets' => $item->retweets,
+                    'impressions' => $item->impressions,
                 ])
             ];
         });
